JIRA:0000- Attended to the feedback on the NHMIS requirement

+++ src/main/java/org/lamisplus/modules/pmtct/domain/entity/PmtctHts.java
@@ -35,7 +35,7 @@ public class PmtctHts {
     private String stageOfPregnancy;
     @Column(name = "patient_uuid")
     private String patientUuid;
-    private Long archived;
+    private Integer archived;
     private String syphilis;
     @Column(name = "hepatitis_b")
     private String hepatitisB;
